pyproject.toml
[tool.poetry]
name = "sql-orm-docker"
version = "0.1.0"
description = ""
authors = ["santhosh <santhosh@gameopedia.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
pydantic = "^2.7.1"
pyyaml = "^6.0.1"
mysql-connector-python = "^8.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
